Perform a /usr-merge on all initrds
Since bookworm, a merged-/usr is mandatory. As we progress in finalizing
the transition, we packages get simplified towards shipping all of their
files below /usr and this will also affect udeb packages. If the
initrds were to stay unmerged, we would expect them to break before too
long. In order to avoid such breakage, /usr-merge them now.

This implementation is similar to the one proposed in
https://salsa.debian.org/installer-team/debootstrap/-/merge_requests/96
and merges after the initial unpack. Merging earlier is prone to
breakage when base-files starts shipping the symlinks. Note that dpkg
--unpack runs preinst scripts and that this is ok, because udebs never
contain preinst scripts.