add '+' as a delimiter for $DATE in $VERSION
I want to be able to add e.g. `+ABI~6.5.0~2` to the version when the
available ABI has been detected (during salsa builds), rather than using
what's been set in the config, but doing so would result in DATE being
set to e.g. `202306XX+ABI~6` because it currently cuts at '.', so this
adds the ability to additionally cut at '+' (and it also saves a shell
invocation by using make functions)