salsa-ci: use the currently available ABI version
This avoids the need to update the configured ABI when testing in
salsa-CI pipelines.

This version of the patch removes the previous (kludgy) introduction of
a 'salsa-ci' build profile.  Instead it uses variables in debian/rules
that can have their default values overridden in order to allow salsa
builds to work.