salsa-CI: if only an unsigned kernel exists, use it
During version skew, the unsigned linux image packages turn up first, so
use them if they're all that is available, just to let CI builds
succeed.

This causes the ABI-check job to fail, so that the user can see
something odd is happening. It also adds '~unsigned' to the generated
artifacts versions, so that one can see the difference.