salsa-ci: add ABI-check (fail when we are skewed)
This adds a test that fails if LINUX_KERNEL_ABI is set, which is what
happens when extract-source detects that we are in version skew, where
the configured ABI is not available in the archive.

This allows the pipeline to produce artifacts (for testing), while still
showing as a failed build, which should ensure that people are not
misled into thinking that they had complete success.