Clarify that partman-auto-recipe.txt is written using BNF
Not everyone is familiar with this notation, and I had an email from
somebody asking for an explanation that suggested they found it
confusing.  The Wikipedia article on BNF is also pretty technical, but
it at least gives people without a formal computer science background a
little more to go on.