extend merge-usr tool to cover base-files functionality
From a DEP17 pov, we're done with the /usr-merge and /usr-move when all
the files installed to canonical locations and no magic scripting is
happening anymore. A recent step towards that goal is that some package
needs to contain the aliasing symbolic links. That package is now
base-files. Unfortunately, there is no base-files udeb, so as more and
more files move, the util/merge-usr will do less and less things and in
particular skip creating the aliasing symbolic links.

Rather than add a base-files udeb for a few symbolic links, I opt for
extending merge-usr to take over this functionality. This change removes
its temporary nature and turns it into a permanent thing though the
merging part could be removed eventually and then it could be renamed to
"alias-usr" or "install-base-files".